+++ CompanyStudio/Store/frmStoreItems.cs
@@ -33,7 +33,7 @@ namespace CompanyStudio.Store
         {
             PermissionsManager.OnLocationPermissionChange += PermissionsManager_OnLocationPermissionChange;
 
-            colQty.ValueType = typeof(short);
+            colQty.ValueType = typeof(decimal);
             colCost.ValueType = typeof(decimal);
 
             LoadData();
@@ -106,7 +106,7 @@ namespace CompanyStudio.Store
         {
             StringBuilder errorBuilder = new StringBuilder();
             DataGridViewRow row = dgvItems.Rows[e.RowIndex];
-            if (row.Cells[colQty.Name].Value is short quantity && quantity < 0)
+            if (row.Cells[colQty.Name].Value is decimal quantity && quantity < 0)
             {
                 errorBuilder.AppendLine("Quantity must be greater than 0");
             }
@@ -125,7 +125,7 @@ namespace CompanyStudio.Store
                 }
 
                 LocationItem item = (LocationItem)otherRow.Tag;
-                if (item.ItemID == thisLocItem.ItemID && (short)row.Cells[colQty.Name].Value == (short)otherRow.Cells[colQty.Name].Value)
+                if (item.ItemID == thisLocItem.ItemID && (decimal)row.Cells[colQty.Name].Value == (decimal)otherRow.Cells[colQty.Name].Value)
                 {
                     errorBuilder.AppendLine("This combination of Item and Quantity already exists");
                     break;
@@ -169,7 +169,7 @@ namespace CompanyStudio.Store
 
                     row.Cells[colImage.Name].Value = item.GetImage();
                     row.Cells[colItem.Name].Value = item.Name;
-                    row.Cells[colQty.Name].Value = (short)0;
+                    row.Cells[colQty.Name].Value = (decimal)0;
                     row.Cells[colCost.Name].Value = 0D;
                     row.Tag = locationItem;
 
@@ -208,7 +208,7 @@ namespace CompanyStudio.Store
                 BasePrice = locationItem.BasePrice,
                 Quantity = locationItem.Quantity
             };
-            itemToSave.Quantity = short.Parse(row.Cells[colQty.Name].Value.ToString());
+            itemToSave.Quantity = decimal.Parse(row.Cells[colQty.Name].Value.ToString());
             itemToSave.BasePrice = decimal.Parse(row.Cells[colCost.Name].Value.ToString());
 
             PutData putData = new PutData(DataAccess.APIs.CompanyStudio, "LocationItem/Put", itemToSave);
@@ -274,8 +274,8 @@ namespace CompanyStudio.Store
                 int compareResult = row1.Cells[itemColName].Value.ToString().CompareTo(row2.Cells[itemColName].Value.ToString());
                 if (compareResult == 0)
                 {
-                    short quantity1 = short.Parse(row1.Cells[quantityColName].Value.ToString());
-                    short quantity2 = short.Parse(row2.Cells[quantityColName].Value.ToString());
+                    decimal quantity1 = decimal.Parse(row1.Cells[quantityColName].Value.ToString());
+                    decimal quantity2 = decimal.Parse(row2.Cells[quantityColName].Value.ToString());
 
                     compareResult = quantity1.CompareTo(quantity2);
                 }
